+++ server/api/emails/marketing/newScreenshots.js
@@ -1,14 +1,17 @@
 const template = require('../template');
 
-module.exports = ({ user, screenshots, unsubscribeLink, userSpaceLink }) => ({
-  subject: `${screenshots.length} nouveaux screenshots ont été ajoutés !`,
-  text: `Bonjour ! ${screenshots.length} nouveaux screenshots ont été ajoutés sur le site. Allez les voir !`,
-  html: template.body(`
+module.exports = ({ user, screenshots, unsubscribeLink, userSpaceLink }) => {
+  const screenshotsAddedText =
+    screenshots.length > 1
+      ? `${screenshots.length} nouveaux screenshots ont été ajoutés`
+      : 'Un nouveau screenshot a été ajouté';
+  return {
+    subject: screenshotsAddedText,
+    text: `Bonjour ! ${screenshotsAddedText} sur le site. Soyez le premier ou la première à les résoudre en vous rendant sur https://guess-the-game.com !`,
+    html: template.body(`
     <div style="text-align: center;">
       ${template.p(`Bonjour ${user.username},`)}
-      ${template.p(
-        `${screenshots.length} nouveaux screenshots ont été ajouté sur guess-the-game.com.`
-      )}
+      ${template.p(`${screenshotsAddedText} sur guess-the-game.com.`)}
       ${template.p(
         'Si vous en connaissez, soyez le premier ou la première à les résoudre en cliquant dessus !'
       )}
@@ -32,4 +35,5 @@ module.exports = ({ user, screenshots, unsubscribeLink, userSpaceLink }) => ({
       `Si vous ne souhaitez plus recevoir ces emails du tout, <a target="_blank" href="${unsubscribeLink}">cliquez sur ce lien.</a>`
     )}
   `),
-});
+  };
+};